update
修改整理类型和合并发送消息需要返回的方法

+++ center/src/main/java/center/client/handle/NotClientLinkHandle.java
@@ -67,7 +67,7 @@ public class NotClientLinkHandle implements Handler {
 	}
 
 	@Override
-	public boolean handler(Sender sender, int clientId, Message message, int mapId, long sequence) {
+	public boolean handler(Sender sender, int clientId, Message message, int mapId, int sequence) {
 		try {
 			ModelProto.NotRegisterClient notification = (ModelProto.NotRegisterClient) message;
 			String clientIp = notification.getCert().toStringUtf8();
