+++ center/src/main/java/center/client/handle/ReqServerInfoHandle.java
@@ -25,7 +25,7 @@ public class ReqServerInfoHandle implements Handler {
 	private static final Logger logger = LoggerFactory.getLogger(ReqServerInfoHandle.class);
 
 	@Override
-	public boolean handler(Sender sender, int clientId, Message message, int mapId, long sequence) {
+	public boolean handler(Sender sender, int clientId, Message message, int mapId, int sequence) {
 		try {
 			ModelProto.ReqServerInfo request = (ModelProto.ReqServerInfo) message;
 			ServerClientManager manager = Center.getInstance().getServerManager();
