+++ game/src/main/java/game/client/handle/ReqRegisterHandle.java
@@ -21,7 +21,7 @@ public class ReqRegisterHandle implements Handler {
 	private static final Logger logger = LoggerFactory.getLogger(ReqRegisterHandle.class);
 
 	@Override
-	public boolean handler(Sender sender, int clientId, Message message, int mapId, long sequence) {
+	public boolean handler(Sender sender, int clientId, Message message, int mapId, int sequence) {
 		try {
 			ModelProto.ReqRegister request = (ModelProto.ReqRegister) message;
 			ModelProto.ServerInfo serverInfo = request.getServerInfo();
@@ -63,7 +63,7 @@ public class ReqRegisterHandle implements Handler {
 	/**
 	 * 发送注册响应
 	 */
-	private void sendRegistrationResponse(Sender sender, int clientId, int mapId, long sequence) {
+	private void sendRegistrationResponse(Sender sender, int clientId, int mapId, int sequence) {
 		ModelProto.AckRegister.Builder response = ModelProto.AckRegister.newBuilder();
 		response.setServerInfo(Game.getInstance().getServerInfo().build());
